# core/server/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Agent:
    properties:
      created:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_ping:
        type: string
      name:
        type: string
      public_ip:
        type: string
      status:
        description: online, offline, unknown
        type: string
      token:
        type: string
    type: object
...
  models.AgentNginxConfig:
    properties:
      agent_id:
        type: string
      domains:
        items:
          $ref: '#/definitions/models.NginxConfig'
        type: array
    type: object
...
  models.CreateDomainRequest:
    properties:
      agent_id:
        type: string
      name:
        example: dashboard.domain.net
        type: string
      ssl_enabled:
        example: false
        type: boolean
    type: object
  models.CreateDomainRouteRequest:
    properties:
      container_name:
        example: my-container
        type: string
      domain_id:
        type: string
      path:
        example: /api
        type: string
      port:
        example: "3000"
        type: string
    type: object
...
  models.Domain:
    properties:
      agent_id:
        description: ID агента, на котором размещен домен
        type: string
      agent_ip:
        description: IP адрес агента
        type: string
      agent_name:
        description: Дополнительные поля для совместимости с frontend
        type: string
      created:
        type: string
      id:
        type: string
      is_active:
        description: активен ли домен
        type: boolean
      name:
        description: 'например: "dashboard.domain.net"'
        type: string
      ssl_enabled:
        description: включен ли SSL
        type: boolean
      updated:
        type: string
    type: object
  models.DomainDetail:
    properties:
      agent:
        $ref: '#/definitions/models.Agent'
      agent_id:
        description: ID агента, на котором размещен домен
        type: string
      agent_ip:
        description: IP адрес агента
        type: string
      agent_name:
        description: Дополнительные поля для совместимости с frontend
        type: string
      created:
        type: string
      id:
        type: string
      is_active:
        description: активен ли домен
        type: boolean
      name:
        description: 'например: "dashboard.domain.net"'
        type: string
      routes:
        items:
          $ref: '#/definitions/models.DomainRoute'
        type: array
      ssl_enabled:
        description: включен ли SSL
        type: boolean
      updated:
        type: string
    type: object
  models.DomainListResponse:
    properties:
      domains:
        items:
          $ref: '#/definitions/models.DomainDetail'
        type: array
      total:
        type: integer
    type: object
  models.DomainRoute:
    properties:
      container_name:
        description: имя контейнера
        type: string
      created:
        type: string
      domain_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      path:
        description: путь (опционально, для /api/*)
        type: string
      port:
        description: порт контейнера
        type: string
      updated:
        type: string
    type: object
  models.DomainRouteListResponse:
    properties:
      routes:
        items:
          $ref: '#/definitions/models.DomainRoute'
        type: array
      total:
        type: integer
    type: object
  models.DomainStatus:
    properties:
      agent_id:
        type: string
      agent_ip:
        type: string
      agent_name:
        type: string
      domain_id:
        type: string
      domain_name:
        type: string
      is_active:
        type: boolean
      routes:
        items:
          $ref: '#/definitions/models.RouteStatus'
        type: array
      ssl_enabled:
        type: boolean
    type: object
...
  models.NginxConfig:
    properties:
      agent_ip:
        type: string
      domain:
        type: string
      routes:
        items:
          $ref: '#/definitions/models.NginxRoute'
        type: array
      ssl_cert:
        $ref: '#/definitions/models.SSLCertificate'
      ssl_enabled:
        type: boolean
    type: object
  models.NginxRoute:
    properties:
      container_name:
        description: имя контейнера
        type: string
      path:
        description: 'например: "/" или "/api"'
        type: string
      port:
        description: порт контейнера
        type: string
    type: object
...
  models.RouteStatus:
    properties:
      container_name:
        type: string
      container_status:
        description: running, stopped, etc.
        type: string
      is_active:
        type: boolean
      path:
        type: string
      port:
        type: string
      route_id:
        type: string
    type: object
  models.SSLCertificate:
    properties:
      private_key:
        type: string
      public_key:
        type: string
    type: object
...
  models.UpdateDomainRequest:
    properties:
      agent_id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      ssl_enabled:
        type: boolean
    type: object
  models.UpdateDomainRouteRequest:
    properties:
      container_name:
        type: string
      is_active:
        type: boolean
      path:
        type: string
      port:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для системы мониторинга Docker контейнеров
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Docker Monitoring System API
  version: "1.0"
paths:
...
  /api/agents/{agent_id}/nginx-config:
    get:
      description: Возвращает конфигурацию nginx для указанного агента
      parameters:
      - description: ID агента
        in: path
        name: agent_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AgentNginxConfig'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Получить конфигурацию nginx агента
      tags:
      - nginx
  /api/domains:
    get:
      description: Возвращает список всех доменов с их маршрутами
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DomainListResponse'
        "500":
          description: Internal Server Error
      summary: Получить список доменов
      tags:
      - domains
    post:
      consumes:
      - application/json
      description: Создает новый домен с указанным агентом
      parameters:
      - description: Данные домена
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/models.CreateDomainRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Domain'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Создать новый домен
      tags:
      - domains
  /api/domains/{domain_id}/routes:
    get:
      description: Возвращает все маршруты для указанного домена
      parameters:
      - description: ID домена
        in: path
        name: domain_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DomainRouteListResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Получить маршруты домена
      tags:
      - domain-routes
  /api/domains/{id}:
    delete:
      description: Удаляет домен и все его маршруты
      parameters:
      - description: ID домена
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Удалить домен
      tags:
      - domains
    get:
      description: Возвращает детальную информацию о домене
      parameters:
      - description: ID домена
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DomainDetail'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Получить домен по ID
      tags:
      - domains
    put:
      consumes:
      - application/json
      description: Обновляет информацию о домене
      parameters:
      - description: ID домена
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Domain'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Обновить домен
      tags:
      - domains
  /api/domains/{id}/status:
    get:
      description: Возвращает статус домена с информацией о контейнерах
      parameters:
      - description: ID домена
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DomainStatus'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Получить статус домена
      tags:
      - domains
  /api/domains/routes:
    post:
      consumes:
      - application/json
      description: Создает новый маршрут для домена
      parameters:
      - description: Данные маршрута
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/models.CreateDomainRouteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DomainRoute'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Создать маршрут домена
      tags:
      - domain-routes
  /api/domains/routes/{id}:
    delete:
      description: Удаляет маршрут домена
      parameters:
      - description: ID маршрута
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Удалить маршрут домена
      tags:
      - domain-routes
    put:
      consumes:
      - application/json
      description: Обновляет информацию о маршруте домена
      parameters:
      - description: ID маршрута
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDomainRouteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DomainRoute'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Обновить маршрут домена
      tags:
      - domain-routes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"